+++ packages/react-ui/storybook/stories/widgetsUI/RangeWidgetUI.stories.js
@@ -1,5 +1,6 @@
 import React from 'react';
-import RangeWidgetUI from '../../../src/widgets/RangeWidgetUI';
+import RangeWidgetUI from '../../../src/widgets/RangeWidgetUI/RangeWidgetUI';
+import { Label, ThinContainer } from '../../utils/storyStyles';
 
 const options = {
   title: 'Organisms/Widgets/RangeWidgetUI',
@@ -19,7 +20,39 @@ const Template = (args) => {
   return <RangeWidgetUI {...args} />;
 };
 
-// TODO
+const LoadingTemplate = (args) => {
+  if (args.series && !Array.isArray(args.series)) {
+    args.series = [];
+  }
+
+  return (
+    <>
+      <Label variant='body1' mb={3}>
+        {'Limited width'}
+      </Label>
+      <ThinContainer>
+        <RangeWidgetUI {...args} />
+      </ThinContainer>
+
+      <Label variant='body1' mt={8} mb={3}>
+        {'Responsive'}
+      </Label>
+      <RangeWidgetUI {...args} />
+    </>
+  );
+};
+
+const data = {
+  data: [400, 500],
+  min: 0,
+  max: 1000,
+  limits: [300, 950]
+};
+
 export const Default = Template.bind({});
-const DefaultProps = { data: [400, 500], min: 0, max: 1000, limits: [300, 950] };
+const DefaultProps = { ...data };
 Default.args = DefaultProps;
+
+export const Loading = LoadingTemplate.bind({});
+const LoadingProps = { ...data, isLoading: true };
+Loading.args = LoadingProps;